requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.115.0
uvicorn>=0.24.0
openai>=1.0.0
numpy>=1.24.0
requests>=2.31.0
pydantic>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0